fix: Enums -> rawValeu em Int16

+++ Sources/SuggestionEngine/models/Intensity.swift
@@ -9,7 +9,7 @@ import Foundation
 // MARK: - Enums
 
 // Intensidades
-public enum Intensity: Int, Comparable, CaseIterable, Sendable {
+public enum Intensity: Int16, Comparable, CaseIterable, Sendable {
     case low = 1, medium = 2, high = 3
     public static func < (lhs: Intensity, rhs: Intensity) -> Bool {
         return lhs.rawValue < rhs.rawValue
